video display sizing fixed
